Making progress in docs
